auto responder activation

+++ satellite-subsystems/demo/src/utils/timeutils.c
@@ -8,7 +8,7 @@
 #include <freertos/FreeRTOS.h>
 #include <freertos/task.h>
 
-void delay_ms(int milliseconds)
+void delay_ms(unsigned long milliseconds)
 {
 	vTaskDelay(milliseconds / portTICK_RATE_MS);
 }
