+++ satellite-subsystems/demo/src/utils/timeutils.h
@@ -8,6 +8,6 @@
 #ifndef UTILS_TIMEUTILS_H_
 #define UTILS_TIMEUTILS_H_
 
-void delay_ms(int milliseconds);
+void delay_ms(unsigned long milliseconds);
 
 #endif /* UTILS_TIMEUTILS_H_ */
